move acquireNonceLockForChainIdKey and map into helper
